+++ package/src/image.ts
@@ -1,0 +1,7 @@
+export {
+    default,
+    default as Image,
+    getImageProps,
+    getImageBlurSvg,
+    type ManifestEntry,
+} from "./image_optimizer/next_image_shim.js";