+++ package/src/image_optimizer/index.ts
@@ -49,9 +49,18 @@ export {
 } from "./run.js";
 
 export {
+    collectUsedImages,
+    findCodeFiles,
+    extractImageReferences,
+    CODE_EXTENSIONS,
+    IGNORED_DIRS,
+} from "./scan_used.js";
+
+export {
     getImageBlurSvg,
 } from "./blur_svg.js";
 
 export {
     type ManifestEntry,
 } from "./next_image_shim.js";
+